Gives Acolytes skills/items related to their God. (#1701)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request
Gives the Acolytes skills related to their patron, they are as follows:
**Pestra:** Gets a blessed needle.
**Dendor:** Gets average farming and weak taming skills and seed knower
trait.
**Malum:** Gets weak in all categories of smithing and the professional
smith trait.
**Noc:** Gets a random language.
**Abyssor:** Gets average fishing.
**Eora:** Gets average music skill and also empath trait.
**Xylix:** Gets skilled music.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game
As devout followers of THEIR god they should have some skills relating
to said gods, also pestra followers not having a needle blessed by
PESTRA while a random non legal doctor has it is downright insane, all
in all open to suggestions about the skill ratios or new addons.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [ ] You tested this on a local server.
- [ ] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->