Inhumen Lair now becomes Lich Spawning Grounds (#1683)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

After consulting with CocoKailey and Aberra, I've decided to resolve the
issue of the Lich not having a proper spawn area. The inhumen lair was
formerly unused and just a sorry camp with a ridiculous helmet. Now it
still retains the helmet, but will become the spawning ground for the
Lich to have a base of operations like other major antags ( Vampire
Manor, Bandit Camp ).

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game
Formerly, you would just spawn in the middle of the Keep if you rolled
Lich. This would in most cases not only put you in danger, but could end
up with you being round removed immediately. Now this will allow the
player to have their own domain to mess around with and spawn in SAFELY.

Images below although there is a current bug with the Centcomm area
making it very dark.


![image](https://github.com/user-attachments/assets/d3e599c7-cb4f-416f-b45d-14b4a800cef8)

![image2](https://github.com/user-attachments/assets/62b5b3a1-5d3d-4b87-ab8e-a971c14ca0fa)

![image3](https://github.com/user-attachments/assets/ff703e25-1a73-4075-8f25-9f70eceb868b)

![image4](https://github.com/user-attachments/assets/5fefe3e8-5f0c-4b66-8655-d0e2a7fa144f)

![image5](https://github.com/user-attachments/assets/99f597e4-9093-4e47-b1c2-fc2df01d105a)

![image6](https://github.com/user-attachments/assets/e811dab3-3891-4384-a167-bf580545609b)

![image](https://github.com/user-attachments/assets/94d079a6-4649-4fc8-9537-4abe965ab85a)



<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

---------

Co-authored-by: CheffieGithub <113442598+CheffieGithub@users.noreply.github.com>