Revert a part of the town elder's nerfs | Add Cooldown to King/Priest's Announcements. (#1716)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request
🤨 
This PR re-add the dream watcher and the town elder's announcement along
with the reduced cooldown. (1 minute)
This PR also add a cooldown to the king and the priest's announcement by
using the same code.
🤷 

![image](https://github.com/user-attachments/assets/4a243388-fb58-4f96-ab80-94186e8830c1)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game
Because Dreamwatcher was nerfed and they were still commented out and as
@Draggeru said it

> We gave the town elder a new toy of COURSE people are going to play
with it. The vocal minority need to sit down. You can revert just the
announcement change as everything else is fine and in line with lore. If
anyone is abusing the announcements yall need to ahelp it and let admins
deal with it.

If you got a problem with it, Go look at how Kings/Priests use their
announcement system before crying about it here, thanks.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->