Adds a new god event - Schism within the Ten (#1748)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

This PR adds a new major god intervention event - Schism within the Ten!

First, requirements for this event are as follows:

- Only on higher pop
- Can happen only later in the round
- Can happen only once
- Divine pantheon storyteller controls the round (except Astrata)
- Astrata must have both less followers AND influence than any other god
from the divine pantheon

What this event does:

1. Picks strongest challenger god in terms of followers from the divine
pantheon (except Astrata).
2. Followers of the challenger get a warning message to prepare for the
upcoming schism.
3. Few minutes later, schism is announced and Astrata's leadeship of the
Ten is cast in doubt. Everyone gets a new temporary ability to choose a
side in the schism, but only tennites actually count in the end, the
other pantheons receive a message about only pretending and using the
schism to their advantage. Everyone can choose a side only 2 times, then
its final. (They can also stay neutral)
4. Struggle begins, as each side tries to get more supporters on their
side by any means necessary. Supporters of each side can see each other
and enemy supporters for the duration of the event on examine.
5. In less than 2 days, the schism is concluded. The side with the most
alive supporters wins! Winning god gets bonus influence and their
supporters receive triumphs depending on which side they were and if
they are an original follower of the god.

Some changes: There will be a status update in the middle of the round
regarding who is leading, and if the challenger wins, one of his
clergymen will be promoted to the role of Vice Priest, with a few of
priest's abilities.

## Why It's Good For The Game

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

Interesting event, more god event variety.

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->