Triumph Unbuy refunds 100% (#1590)
## About The Pull Request

Triumph Unbuy now refunds 100% of triumphs spent instead of 50%.

## Why It's Good For The Game

Losing 10-15 rounds of triumphs, or more depending on how much you can
sleep in a round, sucks. This can be done either by losing the role you
were trying to roll for or just having to refund it last second due to
having to suddenly leave. Neither scenario is a good thing. Aberra
agreed and I finally got around to doing this.

## Pre-Merge Checklist
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.