Adds changes to rosewood that was supposed to be present in the court physician update (#1840)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

This PR primary adds the court physicians office, which is noticeably
missing with the job update

![PR](https://github.com/user-attachments/assets/dba26460-d505-44d7-a4fc-d270ce12b5c2)
This PR also fixes a fucked up roof above the tailors, adds native seeds
to the soilsons, and adds a hermes to the inquisitors hall
## Why It's Good For The Game

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->
Fixes good, job content map support also good

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->