Immigration/exit config based on map (#1850)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request
Allows for a more dynamic arrival and departure, paving the way for
future maps where it would not make sense to go to or from Kingsfield.
The only tangible changes in this PR are with Daftmarsh (departing to
Vanderlin) and Roguetest (departing to Psydonia.)


![RoguetestOrigin2](https://github.com/user-attachments/assets/6ae9f3c7-ffef-4dd8-9ce1-62d2c1bfa5a1)

![RoguetestOrigin](https://github.com/user-attachments/assets/a72caf10-0ccf-4588-9487-375973a96058)

![DaftmarshTest](https://github.com/user-attachments/assets/86064e35-b283-47c5-8034-a01b527abd79)

![RosewoodTest](https://github.com/user-attachments/assets/9d94ed38-469f-432f-9b11-7c82e9d44441)


<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game
Flavour.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

---------

Co-authored-by: AberrantQuerist <crystalseeker04@gmail.com>