Removes Alchemical Potions from Goldface (#1813)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->
Nukes the Alchemical potions from the Goldface. This includes the health
potions, mana potions, and disease cure potions.

## Why It's Good For The Game

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->
More reason to play Apothecary, the Apothecary is always overshadowed in
terms of the Merchant being able to purchase these potions and this PR
aims to integrate the Apothecary to be more useful and needed in normal
play. Also prevents people from just rushing merchant roundstart to get
the potions.

I think this should at least be TMed to see how it plays out.

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->