Captain nerf (#1814)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request
Nerf captain
Wrestling => Expert
Unarmed/Shield => Skilled
Bows => Average

-1 STR/SPD

This change is going to be here until the Bailiff is ready, which will
combine the veteran and captain and change them to be town alligned as
far as its planned right now. This is subject to change.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game
Captain is not chained to the keep like the Royal Knights (Should be)
and is incredibly strong, this often leads to them running around
looking for valids and then killing them without much difficulty.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [ ] You tested this on a local server.
- [ ] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->