Adds Zybantine accent based on Arabic (#1827)
## About The Pull Request

Adds a couple of new lines for a Zybantine based accent using some
arabic words. Some of these needed to be google translated somewhat
because I speak a dialect of it not the original.

## Why It's Good For The Game

More accents that are thematic and fitting in the Universe of Vanderlin.

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->