Daftmarsh patching + locked steward closet. (#1871)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request
-Replaces train with carriage. (Sprite from Azure.)

![image](https://github.com/user-attachments/assets/e07f503f-2f33-40e8-a037-d4768515c689)
(The bush spawning inside has since been fixed.)
-Patches unsmoothened metal pipes. 
-Adds some SCOMs in places that were missing.
(Vanderlin and Daftmarsh:)
-Steward spare keys are now locked in a closet. No more easy theft. 
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game
Improvements, intended features, fixes. 
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

---------

Co-authored-by: AberrantQuerist <crystalseeker04@gmail.com>