Adds Cloth Backpack and Cloth Pouch (#1886)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

Adds a Cloth Backpack and Cloth pouch. 

The cloth backpack functions like a satchel, can only be worn on the
left back slot, and has a 2x3 storage grid. It can be crafted by sewing,
and requires two cloth and three fiber with a sewing craftdiff of one
(weak).

The cloth pouch works like the works like the normal pouch, being able
to be worn on belt slots and the neck, but only has a 1x2 grid and can
only hold pouch authorized items. It can be crafted by sewing, and
requires one cloth and one fiber with a sewing craftdiff of one (weak).

## Why It's Good For The Game

These items add some more crude forms of storage. The cloth backpack has
half the capacity of a satchel, and the cloth pouch has half the
capacity of a pouch. The pouch being a 1x2 means it takes up the same
amount of cells that it contains so there is no storage increase from
having one in a backpack.

The idea behind the cloth backpack is to give the option of a hands free
storage option to those who want to bother with sewing to make one. It
is difficult to get leather and the weak skincrafting skill to make a
satchel, or backpack for the classes who would want some crude storage.

While the pouch can technically be used to increase your personal
storage, my main reason for adding it is in the hopes it can be mapped
in in larger amounts as a disposable item for stewards to hand out
larger sums of coin in a more professional manner. Pouches are cool
because you can throw them and the coins within them won't scatter.

Right now with sewing you can make a rope or cloth belt for a 1x3 grid,
and that's about it for storage, unless you count the half cloak which
you need skilled sewing for a 2x2.

![Github
image](https://github.com/user-attachments/assets/d9df1536-0d3b-4ba7-91a2-f91ebead37a3)


I'll make any changes you guys want, just let me know.

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->